Fix 5-flip_bits.c

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,7 +9,7 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int count = 0, result;
+	unsigned long int count = 0, result;
 
 	result = n ^ m;
